feat: Profile and Update as components
