+++ src/screens/EducatorProfileUpdate.js
@@ -1,0 +1,18 @@
+import '../styles/styles.css';
+import EducatorLayout from '../components/EducatorsLayout';
+import UpdateProfile from '../components/UpdateProfile';
+
+
+const EducatorProfileUpdate = () => {
+
+    return (
+         
+            <EducatorLayout>
+                <UpdateProfile/>
+            </EducatorLayout>
+        
+    )
+  
+  };
+  
+  export default EducatorProfileUpdate;
