--- src/screens/EducatorsProfile.js
@@ -1,13 +1,0 @@
-import React from 'react';
-function EducatorProfile() {
-
-  return (
-   
-    <div>
-
-     Profile
-    </div>
-  );
-}
-
-export default EducatorProfile;
