--- src/screens/EducatorsProileUpdate.js
@@ -1,13 +1,0 @@
-import React from 'react';
-function EducatorProfileUpdate() {
-
-  return (
-   
-    <div>
-
-     Update Profile
-    </div>
-  );
-}
-
-export default EducatorProfileUpdate;
