+++ src/screens/VolunteerProfileUpdate.js
@@ -1,0 +1,18 @@
+import '../styles/styles.css';
+import VolunteerLayout from '../components/VolunteerLayout';
+import UpdateProfile from '../components/UpdateProfile';
+
+
+const VolunteerProfileUpdate = () => {
+
+    return (
+         
+            <VolunteerLayout>
+                <UpdateProfile/>
+            </VolunteerLayout>
+        
+    )
+  
+  };
+  
+  export default VolunteerProfileUpdate;
